A smaller and simpler approach to the sync api
This version of ABLSync puts more responsibilities on the client but in
return is dramatically simpler than the previous versions. By leaving
some of the details of exactly how the syncing is achieved to the
client, this version should support a wider array of application designs
than the previous version, which was more prescriptive.