Use link settings view for control of link state
ABLSync now provides access to a view controller representing a settings
page instead of a function for enabling/disabling Link. There is no
longer a way to programmatically enable and disable, this is controlled
by the user via this settings dialog. The library also now maintains the
enabled state so that it persists across app runs.