ABLSync.h: remove ObjC add tempo/enabled callbacks
New header ABLSyncSettingsViewController with objective-c class
representing the settings view. This leaves ABLSync.h as a pure C header
file with no Apple framework dependencies.

New callback registration functions in ABLSync allow clients to be
notified of changes to the session tempo and changes in Link enabled
state. Since this is now managed by the settings view controller, the
callback allows client apps to observe when the user has enabled or
disabled Link.