Add life-cycle management to LinkHut example app
When the app enters the background, destroy the audio engine along with
the Link instance if it is not playing. This is the desired background
behavior for apps integrating Link. Leaving a Link instance alive while
in background mode while not playing audio will lead to problems for the
user. Either Link will continue to be active in the background for a
long period or it will be suspended, resulting in problems re-connecting
when the app is resumed.